move collection creation to new first draft controller
