+++ app/components/collections/update/settings_form_component.html.erb
@@ -18,7 +18,7 @@
     </div>
   <% end %>
 
-  <%= render Collections::Create::SettingsComponent.new(form: form) %>
+  <%= render FirstDraftCollections::Create::SettingsComponent.new(form: form) %>
 
   <div class="row">
     <div class="col text-end">
