+++ app/components/first_draft_collections/create/details_component.html.erb
