+++ app/components/first_draft_collections/update/form_component.rb
@@ -4,7 +4,7 @@
 module FirstDraftCollections
   module Update
     # The component that renders the form for updating a first draft of a collection.
-    class FormComponent < Collections::Create::FormComponent
+    class FormComponent < FirstDraftCollections::Create::FormComponent
     end
   end
 end
