+++ app/controllers/first_draft_collections_controller.rb
@@ -8,6 +8,39 @@ class FirstDraftCollectionsController < ObjectsController
   before_action :authenticate_user!
   before_action :ensure_sdr_updatable
 
+  def new
+    collection = Collection.new(creator: current_user)
+    authorize! collection
+
+    collection_version = CollectionVersion.new(collection: collection)
+    @form = CreateCollectionForm.new(collection_version: collection_version, collection: collection)
+    @form.prepopulate!
+  end
+
+  # rubocop:disable Metrics/AbcSize
+  # rubocop:disable Metrics/MethodLength
+  def create
+    collection = Collection.new(creator: current_user)
+    authorize! collection
+
+    collection_version = CollectionVersion.new(collection: collection)
+    @form = collection_form(collection_version)
+    if @form.validate(create_params) && @form.save
+      collection_version.collection.event_context = { user: current_user }
+      collection_version.update_metadata!
+      if deposit_button_pushed?
+        collection_version.begin_deposit!
+        redirect_to dashboard_path
+      else
+        redirect_to collection_path(collection)
+      end
+    else
+      render :new, status: :unprocessable_entity
+    end
+  end
+  # rubocop:enable Metrics/MethodLength
+  # rubocop:enable Metrics/AbcSize
+
   def edit
     collection = Collection.find(params[:id])
     authorize! collection
