+++ app/views/first_draft_collections/new.html.erb
@@ -6,7 +6,7 @@
   <div class="container px-4 px-md-3">
     <div id="editor">
       <h1>Create a collection</h1>
-      <%= render Collections::Create::FormComponent.new(collection_form: @form) %>
+      <%= render FirstDraftCollections::Create::FormComponent.new(collection_form: @form) %>
     </div>
   </div>
 </main>
