+++ spec/components/first_draft_collections/create/details_component_spec.rb
@@ -3,7 +3,7 @@
 
 require 'rails_helper'
 
-RSpec.describe Collections::Create::DetailsComponent, type: :component do
+RSpec.describe FirstDraftCollections::Create::DetailsComponent, type: :component do
   pending "add some examples to (or delete) #{__FILE__}"
 
   # it "renders something useful" do
